ISORROPIA op: unify into AerosolExt, drop NonlinearSolve dep, document k_mt·Δt
Polish (no change to the operator's numerical behavior):
- Merge the ISORROPIA operator (was ext/AerosolNonlinearSolveExt.jl) into ext/AerosolExt.jl,
  matching the IsorropiaOp docstring which already pointed at AerosolExt. All GEOS-Chem<->Aerosol
  coupling (het, cloud, ISORROPIA) now lives in one [Aerosol]-keyed extension.
- Drop NonlinearSolve from [weakdeps]/[extensions]: the regime+bisection solve no longer uses
  it. (Kept in [compat]/[extras]/[targets] — still used by climate_forcing/radiation/oh tests.)
  Loading the operator now needs only `using GasChem, Aerosol`.
- Document the k_mt stability constraint in the docstring: k_mt*Dt <= 1, with k_mt = 1/Dt for
  instantaneous equilibrium; reflect the regime+bisection solve.
- Rename the operator test -> test/isorropia_op_test.jl; update the extension reference.

Full suite 1341/1341.
